Split DomainConfigsSection into focused child components
672-line section becomes a 417-line container plus five presentational
children — per-kind row, status badge, expandable field grid,
header/ensure panel, and a guarded-action dialog wrapping the kit's
AdminConfirmDialog — with a small shared types module (script setup
blocks cannot export types). All mutations, the fetch state machine,
and confirm/edit orchestration stay in the parent; children render
props and emit intents.

Zero behavior change: testids, i18n keys, DOM structure, aria
attributes, and classes are preserved verbatim, and the existing spec
passes unmodified. The load-state v-if chain deliberately stays in the
parent — it is the fetch state machine, and splitting its mutually
exclusive conditions across a component boundary risks silent
divergence.
